+++ certbot/CHANGELOG.md
@@ -9,6 +9,9 @@ Certbot adheres to [Semantic Versioning](https://semver.org/).
 * Added `--no-reuse-key`. This remains the default behavior, but the flag may be
   useful to unset the `--reuse-key` option on existing certificates.
 
+* Added `--timeout`. This option specifies the timeout value in seconds
+  for obtaining a certificate from Certificate Authority.
+
 ### Changed
 
 *